repurposed emailfragment to become sign in fragment, added a font, moved around methods to make login work on signinfragment, and deleted login fields from startscreenfragment
