+++ Worldepth/app/src/main/java/com/example/leodw/worldepth/ui/SectionsStatePagerAdapter.java
@@ -4,9 +4,6 @@ import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
 
-import com.example.leodw.worldepth.ui.signup.Email.EmailFragment;
-import com.example.leodw.worldepth.ui.signup.Password.PasswordFragment;
-
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
