+++ Worldepth/app/src/main/java/com/example/leodw/worldepth/ui/StartScreen/StartScreenFragment.java
@@ -2,10 +2,8 @@ package com.example.leodw.worldepth.ui.StartScreen;
 
 import android.arch.lifecycle.ViewModelProviders;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v7.widget.VectorEnabledTintResources;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -14,15 +12,8 @@ import android.view.ViewGroup;
 import com.example.leodw.worldepth.R;
 import com.example.leodw.worldepth.data.FirebaseWrapper;
 import com.example.leodw.worldepth.ui.MainActivity;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
 
 import android.widget.Button;
-import android.widget.EditText;
-import android.widget.ImageView;
 import android.widget.Toast;
 
 import androidx.navigation.NavOptions;
@@ -34,8 +25,6 @@ public class StartScreenFragment extends Fragment {
     private static final String TAG = "StartScreenFragment";
 
     private Button mTestUserLogin;
-    private EditText mEmailInput;
-    private EditText mPasswordInput;
 
     private StartScreenViewModel mViewModel;
     private FirebaseWrapper mFb;
@@ -65,9 +54,7 @@ public class StartScreenFragment extends Fragment {
 
     @Override
     public void onViewCreated(final View view, Bundle savedInstanceState) {
-        createSignupTransitions();
-        mEmailInput = view.findViewById(R.id.enterEmail);
-        mPasswordInput = view.findViewById(R.id.enterPassword);
+        //createSignupTransitions();
         mTestUserLogin = view.findViewById(R.id.testUserLogin);
         mTestUserLogin.setOnClickListener(v -> {
             login("JohnDoe@gmail.com", "Qwerty123!");
@@ -81,18 +68,13 @@ public class StartScreenFragment extends Fragment {
                     mAnimExtras);
         });
 
-        Button logIn = view.findViewById(R.id.loginBtn);
-        logIn.setOnClickListener((view4) -> {
-            if ((!mEmailInput.getText().toString().equals("")) && (!mPasswordInput.getText().toString().equals("")))
-            {
-                login(mEmailInput.getText().toString(), mPasswordInput.getText().toString());
-            } else {
-                Toast.makeText(getContext(), "Enter your login info.", Toast.LENGTH_SHORT).show();
-            }
+        Button toSignIn = view.findViewById(R.id.toSignIn);
+        toSignIn.setOnClickListener((view4) -> {
+            Navigation.findNavController(getView()).navigate(R.id.action_startScreenFragment_to_SignInFragment);
         });
     }
 
-    private void createSignupTransitions() {
+    /*private void createSignupTransitions() {
         ImageView logo = getView().findViewById(R.id.start_logo);
         mAnimExtras = new FragmentNavigator.Extras.Builder()
                 .addSharedElement(logo, "bigLogo")
@@ -100,7 +82,7 @@ public class StartScreenFragment extends Fragment {
         mNavOptions = new NavOptions.Builder()
                 .setEnterAnim(R.animator.signup_anim)
                 .build();
-    }
+    }*/
 
     private void login(String email, String password) {
         mFb.getFirebaseAuth().signInWithEmailAndPassword(email, password).
