--- Worldepth/app/src/main/java/com/example/leodw/worldepth/ui/signup/Email/EmailViewModel.java
@@ -1,6 +1,0 @@
-package com.example.leodw.worldepth.ui.signup.Email;
-
-import android.arch.lifecycle.ViewModel;
-
-public class EmailViewModel extends ViewModel {
-}
