+++ Worldepth/app/src/main/java/com/example/leodw/worldepth/ui/signup/SignIn/SignInViewModel.java
@@ -1,0 +1,6 @@
+package com.example.leodw.worldepth.ui.signup.SignIn;
+
+import android.arch.lifecycle.ViewModel;
+
+public class SignInViewModel extends ViewModel {
+}
