+++ Worldepth/app/src/main/res/layout/fragment_location.xml
@@ -68,7 +68,6 @@
         android:id="@+id/exploreButton"
         android:layout_width="200dp"
         android:layout_height="200dp"
-        android:background="null"
         android:src="@drawable/search_icon"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
