+++ Worldepth/app/src/main/res/values/colors.xml
@@ -5,7 +5,7 @@
     <color name="colorAccent">#D81B60</color>
     <color name="colorPrimaryLight">#5F71B5</color>
     <color name="colorWindowBackground">#333</color>
-    <color name="inputColor">#99b7e8</color>
+    <color name="inputColor">#DDDDDD</color>
     <color name="backgroundColor">#274060</color>
     <color name="black">#000000</color>
     <color name="purple">#8A56AC</color>
