+++ Worldepth/app/src/main/res/values/strings.xml
@@ -49,6 +49,8 @@
     <string name="continue_with">Continue with:</string>
     <string name="general">General</string>
     <string name="me">Me</string>
+    <string name="SIGNIN">SIGN IN</string>
+    <string name="CONTINUE">CONTINUE</string>
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
