+++ Worldepth/app/src/main/res/values/styles.xml
@@ -21,4 +21,7 @@
     <style name="montserrat_b" parent="@android:style/TextAppearance.Small">
         <item name="android:fontFamily">@font/montserrat_b</item>
     </style>
+    <style name="montserrat_m" parent="@android:style/TextAppearance.Small">
+        <item name="android:fontFamily">@font/montserrat_m</item>
+    </style>
 </resources>
